backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
pydantic-settings>=2.0.0
